+++ src/mocks/handlers.ts
@@ -1,17 +1,13 @@
-import { rest } from "msw";
+import { HttpResponse, http } from "msw";
 const apng =
   "iVBORw0KGgoAAAANSUhEUgAAAAUAAAAFCAMAAAC6sdbXAAAACGFjVEwAAAADAAAAAM7tusAAAAAMUExURQMDAwAAAJKSkv8AAGIb5p4AAAABdFJOUwBA5thmAAAAGmZjVEwAAAAAAAAABQAAAAUAAAAAAAAAAAAUAGQAANMipokAAAAOSURBVAjXY2AEAQYcJAABlQAaPUDJKQAAABpmY1RMAAAAAQAAAAUAAAAFAAAAAAAAAAAAFABkAABIUUxdAAAAEmZkQVQAAAACCNdjYAYBBhwkAASDAEy9AgcTAAAAGmZjVEwAAAADAAAABQAAAAUAAAAAAAAAAAAUAGQAAKXHn7QAAAASZmRBVAAAAAQI12NgAgEGHCQAAwwAMyEXezIAAAAbdEVYdFNvZnR3YXJlAEFQTkcgQXNzZW1ibGVyIDIuN8Hj04gAAAAASUVORK5CYII=";
 
 export const handlers = [
-  rest.get("http://localhost/image", (req, res, ctx) => {
+  http.get("http://localhost/image", () => {
     const buffer = Buffer.from(apng, "base64");
 
-    return res(
-      ctx.status(200),
-      ctx.set("Content-Type", "image/png"),
-      ctx.body(buffer),
-      ctx.delay(0),
-      ctx.set("Content-Length", buffer.length.toString()),
-    );
+    return HttpResponse.arrayBuffer(buffer, {
+      headers: { "Content-Type": "image/png" },
+    });
   }),
 ];